Test and Tuning in RingBuffer, experience on embedded platform
* HandlePtr32Union_emC no more used, instead
* HandleADDR_emC: defined in compl_adaption.h platform/application-depending.
  It may be the better approach, to decide, an address (pointer) can be used as handle
  for specific platforms (Simulink), but elsewhere it is a really pointer.
  Either, or. A unit cannot be used if it is used as type. Docu should be still done.
* Base/Atomic_emC.h vs. os_AtomicAccess.h: The Atomic_emC.h is the new one.
  As base routines only compareAndSwap should be used and defined os-specific,
  maybe as inline in the compl_adaption.h (for simple platforms without CMPXCHG).
  Hence the conditional (#ifdef) given prototypes are defined there.
  The os_AtomicAccess.h will be no more necessary.
* ExcThreadCxt_emC.c: it is the renamed ThreadContextUserBuffer_emC.c in emC_srcApplSpec
  It is part of the stable emC/Base, but the content is conditionally compiled,
  excluded on DEF_ThreadContext_SIMPLE.
  The ThreadContext for simple interrupt systems can be defined
  in the specific startup sources of that applications.
* Base/EventQu_emC.* in progress, not ready yet.
* Windws/os_thread.c cleanup

+++ emC/Base/ObjectSimple_emC.h
@@ -121,7 +121,7 @@ const Initialization                         */
 
 /**Initializing of a simple object.  */
 #ifdef DEF_ObjectJc_SIMPLE
-  #define INIZ_ObjectJc(OBJ, REFL, ID)  { ((((uint32)(ID_##REFL))<<kBitInstanceType_ObjectJc) & mInstanceType_ObjectJc)  | (sizeof(OBJ) & mSize_ObjectJc) }
+  #define INIZ_ObjectJc(OBJ, REFL, ID)  { ((((uint32)(ID_##REFL))<<kBitInstanceType_ObjectJc) & mInstanceType_ObjectJc)  | (uint32)(sizeof(OBJ) & mSize_ObjectJc) }
 #elif !defined(DEF_ObjectJcpp_REFLECTION)  //for that, definition in Object_emC.h
   #define INIZ_ObjectJc(OBJ, REFL, ID)  \
    { ( (((uint32)(ID))<<kBitIdentSmall_ObjectJc) \
